+++ source/platforms/GitHub.ts
@@ -37,6 +37,7 @@ export const GitHub = (api: GitHubAPI) => {
     name: "GitHub",
 
     api,
+    getReviewInfo: api.getPullRequestInfo,
     getPlatformGitRepresentation: () => gitDSLForGitHub(api),
 
     getPlatformDSLRepresentation: async () => {