+++ CHANGELOG.md
@@ -15,6 +15,14 @@
 
 * Adds support for the GH Checks API [notes to come][@orta][]
 * JSON diffs use the JSON5 parser, so can now ignore comments in JSON etc [@orta][]
+* Allows the synchronous execution of multiple dangerfiles in one single "danger run".
+
+  Not a particularly useful feature for Danger-JS, but it means Peril can combine many runs into a single execution
+  unit. This means people only get 1 message. [@orta][]
+
+# 3.6.6
+
+* Updates vm2 to be an npm published version [@orta][]
 
 # 3.6.5
 
@@ -184,6 +192,7 @@
 * Improvements to the Flow definition file. [@orta][]
 * Improve path generator for danger-runner. [@Mifi][]
 * Update the PR DSL to include bots. [@orta][]
+* Add utility function to build tables in Markdown [@keplersj][]
 
 ## 3.1.7
 
@@ -1076,6 +1085,7 @@ Not usable for others, only stubs of classes etc. - [@orta][]
 [@wizardishungry]: https://github.com/wizardishungry
 [@hongrich]: https://github.com/hongrich
 [@peterjgrainger]: https://github.com/peterjgrainger
+[@keplersj]: https://github.com/keplersj
 [@azz]: https://github.com/azz
 [@mifi]: https://github.com/ionutmiftode
 [@sunshinejr]: https://github.com/sunshinejr